Determine MIME type of uploaded documents on server #36
Rather than rely on value provided by browser, determine the MIME type on the server side.

This change temporarily makes mime_type optional in the client request, to save the existing code from causing a validation error, but this can be removed once it is no longer part of the request.