Fix date part of sample search #73
Because the dates are persisted in text fields in the format d/m/Y,
MySQL can't compare them in a way that makes sense to humans.

As a workaround, the date is transformed to Y-m-d when it is passed into
the object and then transformed back in the getter, for display.